core/pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["setuptools", "poetry-core>=1.0.0"]

[tool.poetry]
authors = ["Côme Arvis <come.arvis@neuronest.net>", "Guillaume Ligner <guillaume.ligner@neuronest.net>"]
description = "Neuronest core project"
name = "neuronest"
packages = [{ include = "neuronest"}]
version = "0.1.12"

[tool.poetry.dependencies]
ffmpeg-python = "^0.2.0"
google-cloud-aiplatform = "^1.18.2"
google-cloud-bigquery = "^2.34.4"
google-cloud-storage = "^2.5.0"
librosa = "^0.9.2"
moviepy = "^1.0.3"
numpy = "^1.23.3"
opencv-contrib-python = "4.5.5.64"
opencv-python = "4.5.5.64"
opencv-python-headless = "4.5.5.64"
pandas = "^1.5.0"
pydantic = "^1.10.2"
python = ">=3.8,<3.11"
tqdm = "^4.64.1"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
black = "^22.8.0"
flake8 = "^5.0.4"
isort = "^5.10.1"
pre-commit = "^2.20.0"
pylint = "^2.15.2"
pytest = "^7.1.3"
pytest-xdist = "^2.5.0"
toml-sort = "^0.20.1"
